Added java othello development project
Developed a fully functional othello-game player, with random, greedy and human strategy capabilities and an analysis of the probabilities and likelihoods of one of two random-brained computer players winning a game against each other. Used JUnit testing for a testing suite, JavaDoc, and multiple object-oriented concepts to create an architecturally sound software design.